requirements.txt
certifi==2018.10.15
chardet==3.0.4
Click==7.0
cov-core==1.15.0
coverage==4.5.1
Flask==1.0.2
Flask-SQLAlchemy==2.3.2
idna==2.7
itsdangerous==0.24
Jinja2==2.10
MarkupSafe==1.0
nose==1.3.7
nose-cov==1.6
PyHamcrest==1.9.0
requests==2.20.0
six==1.11.0
SQLAlchemy==1.2.12
urllib3==1.24
Werkzeug==0.14.1
